feat: implement sync logic

+++ internal/manager/storage/storage.go
@@ -3,23 +3,55 @@ package storage
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/nekr0z/gk/internal/manager/crypt"
 	"github.com/nekr0z/gk/internal/manager/secret"
 )
 
+var ErrNotFound = fmt.Errorf("not found")
+
 // Repository stores secrets.
 type Repository struct {
 	storage    Storage
+	remote     Remote
+	resolver   ResolverFunc
 	passPhrase string
 }
 
 // New creates a new repository.
-func New(storage Storage, passPhrase string) *Repository {
-	return &Repository{
+func New(storage Storage, passPhrase string, opts ...Option) (*Repository, error) {
+	if storage == nil {
+		return nil, errors.New("storage is nil")
+	}
+
+	r := &Repository{
 		storage:    storage,
 		passPhrase: passPhrase,
+	}
+
+	for _, opt := range opts {
+		opt(r)
+	}
+
+	return r, nil
+}
+
+// Option is a function that configures a repository.
+type Option func(*Repository)
+
+// UseRemote sets the remote storage.
+func UseRemote(remote Remote) Option {
+	return func(r *Repository) {
+		r.remote = remote
+	}
+}
+
+// UseResolver sets the conflict resolver.
+func UseResolver(resolver ResolverFunc) Option {
+	return func(r *Repository) {
+		r.resolver = resolver
 	}
 }
 
@@ -50,6 +82,10 @@ func (r *Repository) Read(ctx context.Context, key string) (secret.Secret, error
 		return secret.Secret{}, err
 	}
 
+	if len(storedSecret.EncryptedPayload.Data) == 0 && storedSecret.LastKnownServerHash == [32]byte{} {
+		return secret.Secret{}, fmt.Errorf("secret not found")
+	}
+
 	payload, err := crypt.Decrypt(storedSecret.EncryptedPayload, r.passPhrase)
 	if err != nil {
 		return secret.Secret{}, fmt.Errorf("failed to decrypt secret: %w; is the passphrase correct?", err)
@@ -64,7 +100,39 @@ func (r *Repository) Delete(ctx context.Context, key string) error {
 		return ctx.Err()
 	}
 
-	return r.storage.Delete(ctx, key)
+	current, err := r.storage.Get(ctx, key)
+	if err != nil {
+		if errors.Is(err, ErrNotFound) {
+			// already deleted
+			return nil
+		}
+		return err
+	}
+
+	if current.LastKnownServerHash == [32]byte{} {
+		return r.storage.Delete(ctx, key)
+	}
+
+	current.EncryptedPayload = crypt.Data{}
+	return r.storage.Put(ctx, key, current)
+}
+
+// Sync syncs the key with the remote.
+func (r *Repository) Sync(ctx context.Context, key string) error {
+	if r.remote == nil {
+		return fmt.Errorf("remote storage is not set")
+	}
+
+	return sync(ctx, r.storage, r.remote, r.resolver, key)
+}
+
+// SyncAll syncs all keys with the remote.
+func (r *Repository) SyncAll(ctx context.Context) error {
+	if r.remote == nil {
+		return fmt.Errorf("remote storage is not set")
+	}
+
+	return syncAll(ctx, r.storage, r.remote, r.resolver)
 }
 
 // Storage is a secrets storage.
@@ -72,10 +140,17 @@ type Storage interface {
 	Get(context.Context, string) (StoredSecret, error)
 	Put(context.Context, string, StoredSecret) error
 	Delete(context.Context, string) error
+	List(context.Context) (map[string]ListedSecret, error)
 }
 
 // StoredSecret is a secret encrypted and stored.
 type StoredSecret struct {
 	EncryptedPayload    crypt.Data
+	LastKnownServerHash [32]byte
+}
+
+// ListedSecret is a secret in list.
+type ListedSecret struct {
+	Hash                [32]byte
 	LastKnownServerHash [32]byte
 }
